systemc: Add a filter file for the tests which are expected to work.
Not all of the tests this filter accepts actually work, but we haven't
yet explicitly decided not to try to make them tests work yet.

Recording this in a file makes test runs more repeateable, and creates
a place to record why certain tests or groups of tests are being
excluded.

Change-Id: I91d44b6500a3e8ff5d5808222ce03f138374cf8b
Reviewed-on: https://gem5-review.googlesource.com/11258
Reviewed-by: Gabe Black <gabeblack@google.com>
Maintainer: Gabe Black <gabeblack@google.com>

Former-commit-id: b29a2d00b85f4c51b9abe6651a8145a0e1eb5484