misc: Fix BaseCPU doxygen
Doxygen was stopping with #include "arch/null/cpu_dummy.hh" so the html
for BaseCPU was only the three functions defined in the dummy CPU. This
forces doxygen to skip this #include correctly.

Note: The file references for base_cpu still aren't quite right, but it's
better than it was.

Change-Id: Ifafe247df2511caee2569d534bd29348a5ce9e8e
Signed-off-by: Jason Lowe-Power <jason@lowepower.com>
Reviewed-on: https://gem5-review.googlesource.com/11649
Reviewed-by: Brandon Potter <Brandon.Potter@amd.com>

Former-commit-id: c9dd86a5cd6f26a2ba776d0adb60d4e18c553b8b